% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenyacensus.R
\name{get_census_year}
\alias{get_census_year}
\title{Get Census Data by Year}
\usage{
get_census_year(years = "all")
}
\arguments{
\item{years}{Numeric vector of years. Use "all" for all years.}
}
\value{
A data frame with census data for selected years
}
\description{
Retrieve national census data for specific years.
}
\examples{
get_census_year(2019)
get_census_year(c(1969, 1979, 1989, 1999, 2009, 2019))
}
