% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenyacensus.R
\name{get_county_data}
\alias{get_county_data}
\title{Get County-Level Data}
\usage{
get_county_data(counties = "all", variables = NULL)
}
\arguments{
\item{counties}{Character vector of county names. Use "all" for all counties.}

\item{variables}{Character vector of variables to return. Default is all.}
}
\value{
A data frame with selected counties and variables
}
\description{
Retrieve 2019 census data for specific counties.
}
\examples{
get_county_data("Nairobi City")
get_county_data(c("Nairobi City", "Mombasa", "Kisumu"))
get_county_data("all", c("County", "Total", "Population_Density"))
}
