% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenyacensus.R
\docType{package}
\name{kenyacensus}
\alias{kenyacensus}
\alias{_PACKAGE}
\alias{kenyacensus-package}
\title{Kenya Census Data Package}
\description{
A comprehensive collection of Kenya Population and Housing Census data
from 1948 to 2019, sourced from the Kenya National Bureau of Statistics (KNBS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/musanyaks/kenyacensus}
  \item Report bugs at \url{https://github.com/musanyaks/kenyacensus/issues}
}

}
\author{
\strong{Maintainer}: Musa Rioba \email{nyakerabachi@gmail.com}

}
\keyword{internal}
