% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenyacensus.R
\name{plot_population_pyramid}
\alias{plot_population_pyramid}
\title{Plot Population Pyramid}
\usage{
plot_population_pyramid()
}
\value{
A ggplot object
}
\description{
Creates a population pyramid showing age and sex distribution.
}
\examples{
plot_population_pyramid()
}
